requirements.txt
discord.py
aiosqlite
pynaclcls